Transition alerts to Wagtail
